move typecasting from binary comp. trait for better BT
